# docs/product/case-workspace/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Consultify — Case Workspace API
  version: 1.0.0-as-built-2026-08-10
  description: |
    HTTP contract for the 11 Case Workspace domain services.

    **Envelope.** Every success response is `{ "data": <payload> }`. The one
    exception is `GET /process-definitions`, which is the only natively
    paginated list in the codebase and returns
    `{ "data": [...], "nextCursor": string|null }`.

    **Errors.** Every error response is produced by
    `server/src/utils/ErrorHandler.ts#errorHandlerMiddleware` and is shaped
    `{ "status": "fail"|"error", "correlationId": string|null,
       "error": { "code": string, "message": string, ... } }`.
    `error.code` is the SCREAMING_SNAKE upper-casing of the domain service's
    own `snake_case` error code (`_shared/errors.ts#toCaseWorkspaceAppError`),
    so it is stable across releases and safe to branch on.

    **Correlation.** An inbound `X-Correlation-ID` request header is honoured
    and echoed; when absent, one is minted per request
    (`_shared/errors.ts#resolveCorrelationId`).

    **Optimistic concurrency.** Mutations on versioned aggregates take
    `expectedVersion` in the request BODY (not an `If-Match` header). A stale
    value yields `409` with a `*_VERSION_CONFLICT` code and no mutation.

    **Idempotency.** Where a command has an idempotency primitive, the route
    reads the `Idempotency-Key` request header, falling back to a body field.
...
servers:
  - url: /api/v8/case-workspace
...
security:
  - bearerAuth: []
...
paths:
...
  /cases:
    post:
      tags: [Cases]
      operationId: createCase
      summary: Create a Case.
      description: |
        `organizationId` is forced to the authenticated org; `createdByActorId`
        to the authenticated user. `case_core.project_id` is UNIQUE, so a
        second Case for the same project fails with
        `409 CASE_ALREADY_EXISTS_FOR_PROJECT`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCaseBody' }
      responses:
        '201': { $ref: '#/components/responses/CaseCreated' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/OrgRoleDenied' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [Cases]
      operationId: listCases
      summary: List Cases in the authenticated organization.
      description: |
        Not paginated — the service returns a plain ordered array. There is no
        `cursor`/`limit` parameter on this route today (CW-DOD-I5 gap).
      parameters:
        - { in: query, name: caseStatus, required: false, schema: { $ref: '#/components/schemas/CaseStatus' } }
        - { in: query, name: caseProfile, required: false, schema: { $ref: '#/components/schemas/CaseProfile' } }
        - { in: query, name: governanceTier, required: false, schema: { $ref: '#/components/schemas/GovernanceTier' } }
      responses:
        '200':
          description: Case list.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CaseCoreView' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/OrgRoleDenied' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/by-project/{projectId}:
    get:
      tags: [Cases]
      operationId: getCaseByProject
      summary: Fetch the single Case bound to a project.
      description: |
        Enumeration-safe: a Case belonging to another organization is reported
        as `404 CASE_NOT_FOUND`, identically to a projectId that has no Case.
      parameters:
        - { in: path, name: projectId, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [Cases]
      operationId: getCase
      summary: Fetch one Case.
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/status:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: transitionCaseStatus
      summary: Transition Case status.
      description: |
        Illegal transitions fail `409 CASE_STATUS_TRANSITION_NOT_ALLOWED`
        (the raw message carries a `:FROM->TO` suffix; `error.code` does not).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetStatus]
              properties:
                targetStatus: { $ref: '#/components/schemas/CaseStatus' }
                reason: { type: string, minLength: 1 }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/governance-tier:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: updateCaseGovernanceTier
      summary: Change the governance tier (append-only history).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier, reason]
              properties:
                tier: { $ref: '#/components/schemas/GovernanceTier' }
                reason: { type: string, minLength: 1 }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/autonomy-policy:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: updateCaseAutonomyPolicy
      summary: Change the Case autonomy policy, capped by the org maximum.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy]
              properties:
                policy: { $ref: '#/components/schemas/AutonomyPolicy' }
                orgMaxAutonomyPolicy:
                  allOf: [{ $ref: '#/components/schemas/AutonomyPolicy' }]
                  nullable: true
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/closure-axis:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: updateCaseClosureAxisStatus
      summary: Set one of the four closure-axis statuses.
      description: |
        The route's zod enum accepts all four values
        (`NOT_APPLICABLE|PENDING|COMPLETED|VALIDATED`), but the service narrows
        per axis: `delivery|decision|implementation` accept `COMPLETED` and
        reject `VALIDATED`; `outcome` accepts `VALIDATED` and rejects
        `COMPLETED`. A mismatched pair is rejected by the service, not by zod.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [axis, status]
              properties:
                axis: { type: string, enum: [delivery, decision, implementation, outcome] }
                status: { type: string, enum: [NOT_APPLICABLE, PENDING, COMPLETED, VALIDATED] }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/closure:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: recordCaseClosure
      summary: Record the contracted closure type.
      description: |
        Records `closure_type`; it does NOT set `case_status='CLOSED'` (that is
        `POST /cases/{caseId}/status`). A second call fails
        `409 CASE_CLOSURE_ALREADY_RECORDED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [closureType]
              properties:
                closureType: { $ref: '#/components/schemas/ClosureType' }
                evidenceRef: { type: string, minLength: 1, nullable: true }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /cases/{caseId}/cancel:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    post:
      tags: [Cases]
      operationId: cancelCase
      summary: Cancel a Case (thin wrapper over the CANCELLED status transition).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1 }
      responses:
        '200': { $ref: '#/components/responses/CaseOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ===========================================================================
  # casePlanVersions.routes.ts — casePlanVersionService
  # ===========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Verified by `verifyToken` and `requireV8OrgContext` upstream of this
        router (routes/v8/index.ts:55-56, 83). The case-workspace handlers
        themselves never parse a token — they read the already-populated
        `req.v8Context`.
  parameters:
    CaseId:
      { in: path, name: caseId, required: true, schema: { type: string, minLength: 1 } }
...
  responses:
...
    ValidationError:
      description: |
        Request body / query / path parameters failed zod validation, or a
        required idempotency key was absent. `error.code` is
        `VALIDATION_ERROR`; `error.issues[]` carries `{path,message,code}` per
        failure (see `_shared/validate.ts`). Some routes emit a bare
        `{ error: { code, message } }` for the missing-idempotency-key case
        without going through the error middleware.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    OrgRoleDenied:
      description: |
        `not_org_member` or `insufficient_org_role` (403), or a genuine
        caller-identity authorization failure (`SELF_APPROVAL_FORBIDDEN`,
        `PLAN_DIFF_CROSS_CASE_FORBIDDEN`,
        `PROCESS_DEFINITION_SHARE_NOT_AUTHORIZED`). NOT used for cross-tenant
        resource access — that is 404.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: |
        The resource does not exist, OR it exists in a different tenant
        (`case_access_denied` → 404, SEC-009 enumeration safety).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: |
        Optimistic-concurrency conflict (`*_VERSION_CONFLICT`), uniqueness
        conflict (`*_ALREADY_*`, `IDEMPOTENCY_KEY_CONFLICT`), or the resource
        is in the wrong lifecycle state for this action (`*_NOT_EDITABLE`,
        `*_NOT_ACTIVE`, `*_EXPIRED`, `*_VALIDATION_FAILED`,
        `*_STATUS_TRANSITION_NOT_ALLOWED`). No mutation occurred.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unprocessable:
      description: |
        Cross-reference / cross-tenant consistency failure between
        caller-supplied ids (`*_MISMATCH`, `*_INVALID`), or a business rule
        this codebase classifies as unprocessable. Also the fallback status for
        any unrecognized `snake_case` domain error code.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InternalError:
      description: |
        Unexpected failure, or a genuine internal invariant breach
        (`*_INSERT_FAILED`, `*_IDEMPOTENCY_RECORD_NOT_FOUND`). Never leaks a
        stack trace or SQL text in production.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    CaseOk:
      description: Case readback.
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: '#/components/schemas/CaseCore' }
    CaseCreated:
      description: Case created.
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: { $ref: '#/components/schemas/CaseCore' }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        status: { type: string, enum: [fail, error] }
        correlationId: { type: string, nullable: true }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                SCREAMING_SNAKE upper-casing of the domain service's own
                snake_case code, or `VALIDATION_ERROR` / `INTERNAL_ERROR`.
            message: { type: string }
            issues:
              type: array
              description: Present on VALIDATION_ERROR only.
              items:
                type: object
                properties:
                  path: { type: string }
                  message: { type: string }
                  code: { type: string }
          additionalProperties: true

    # -- shared request bodies --------------------------------------------------
...
    CaseProfile: { type: string, enum: [LIGHT, STANDARD, TRANSFORMATION, MONITORING] }
    GovernanceTier: { type: string, enum: [LIGHTWEIGHT, STANDARD, CONTROLLED] }
    AutonomyPolicy: { type: string, enum: [ASK_EACH_ACTION, ASK_MATERIAL_ACTIONS, EXECUTE_APPROVED_PLAN] }
    CaseStatus: { type: string, enum: [DRAFT, ACTIVE, BLOCKED, CLOSED, FAILED, CANCELLED] }
    ClosureType:
      type: string
      enum: [DELIVERY_COMPLETED, DECISION_COMPLETED, IMPLEMENTATION_COMPLETED, OUTCOME_VALIDATED, COMPLETED_PARTIAL]
    ClosureAxisStatus: { type: string, enum: [NOT_APPLICABLE, PENDING, COMPLETED] }
    OutcomeAxisStatus: { type: string, enum: [NOT_APPLICABLE, PENDING, VALIDATED] }
...
    CreateCaseBody:
      type: object
      required: [projectId, contractedClosureType]
      properties:
        projectId: { type: string, minLength: 1 }
        caseProfile: { $ref: '#/components/schemas/CaseProfile' }
        governanceTier: { $ref: '#/components/schemas/GovernanceTier' }
        autonomyPolicy: { $ref: '#/components/schemas/AutonomyPolicy' }
        contractedClosureType: { $ref: '#/components/schemas/ClosureType' }
        sponsorUserId: { type: string, minLength: 1, nullable: true }
        acceptanceCriteriaRef: { type: string, minLength: 1, nullable: true }
      additionalProperties: false
    GovernanceTierHistoryEntry:
      type: object
      required: [tier, changedAt, changedByActorId, reason]
      properties:
        tier: { $ref: '#/components/schemas/GovernanceTier' }
        changedAt: { type: string }
        changedByActorId: { type: string }
        reason: { type: string }
    CaseCore:
      type: object
      required:
        [caseId, projectId, organizationId, caseProfile, governanceTier, governanceTierHistory,
         autonomyPolicy, caseStatus, contractedClosureType, deliveryStatus, decisionStatus,
         implementationStatus, outcomeStatus, createdByActorId, version, createdAt, updatedAt]
      properties:
        caseId: { type: string }
        projectId: { type: string }
        organizationId: { type: string }
        caseProfile: { $ref: '#/components/schemas/CaseProfile' }
        governanceTier: { $ref: '#/components/schemas/GovernanceTier' }
        governanceTierHistory:
          type: array
          items: { $ref: '#/components/schemas/GovernanceTierHistoryEntry' }
        autonomyPolicy: { $ref: '#/components/schemas/AutonomyPolicy' }
        autonomyPolicyRef: { type: string, nullable: true }
        caseStatus: { $ref: '#/components/schemas/CaseStatus' }
        contractedClosureType: { $ref: '#/components/schemas/ClosureType' }
        deliveryStatus: { $ref: '#/components/schemas/ClosureAxisStatus' }
        decisionStatus: { $ref: '#/components/schemas/ClosureAxisStatus' }
        implementationStatus: { $ref: '#/components/schemas/ClosureAxisStatus' }
        outcomeStatus: { $ref: '#/components/schemas/OutcomeAxisStatus' }
        closureType:
          allOf: [{ $ref: '#/components/schemas/ClosureType' }]
          nullable: true
        closedAt: { type: string, nullable: true }
        closedByActorId: { type: string, nullable: true }
        closureEvidenceRef: { type: string, nullable: true }
        sponsorUserId: { type: string, nullable: true }
        acceptanceCriteriaRef: { type: string, nullable: true }
        budgetPolicyRef: { type: string, nullable: true }
        currentPlanVersionId: { type: string, nullable: true }
        createdByActorId: { type: string }
        version: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
        completedAt: { type: string, nullable: true }
    CaseCoreView:
      allOf:
        - $ref: '#/components/schemas/CaseCore'
        - type: object
          description: |
            List projection. `projectName`/`projectDescription`/`projectOwnerId`
            are read-only display fields projected from `projects`.
          properties:
            projectName: { type: string, nullable: true }
            projectDescription: { type: string, nullable: true }
            projectOwnerId: { type: string, nullable: true }

    # -- Plan version -----------------------------------------------------------
...